pyproject.toml
[tool.poetry]
name = "iwork"
version = "0.1.0"
description = ""
authors = ["Askhat <askhat.ss23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.115.11"}
sqlalchemy = "^2.0.39"
pydantic = "^2.10.6"
passlib = "^1.7.4"
python-jose = "^3.4.0"
python-multipart = "^0.0.20"
alembic = "^1.15.1"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
upstash-redis = "^1.3.0"
bcrypt = "^4.3.0"
fastapi-mail = "^1.4.2"
jinja2 = "^3.1.6"
boto3 = "^1.37.17"
authlib = "^1.5.1"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"
yfinance = "^0.2.55"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
uvicorn = "^0.34.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"
sqlalchemy-utils = "^0.41.2"
faker = "^37.1.0"
factory-boy = "^3.3.3"
pytest-env = "^1.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "migrations/*", "alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
